# astroneko-backend/docs/swagger.yaml
basePath: /
definitions:
...
  shared.ResponseBody:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
  shared.Status:
    properties:
      code:
        type: string
      errorID:
        type: string
      message:
        items:
          type: string
        type: array
      successID:
        type: string
    type: object
...
  user.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/user.UserResponse'
    type: object
  user.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  user.CreateUserResponse:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
...
  user.GoogleLoginRequest:
    properties:
      id_token:
        type: string
      refresh_token:
        type: string
    required:
    - id_token
    type: object
...
  user.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  user.RefreshTokenResponse:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
...
  user.UserResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      firebase_uid:
        type: string
      id:
        type: string
      is_activated_referral:
        type: boolean
      latest_login_at:
        type: string
      profile_image_url:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Astroneko Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Astroneko Backend API
  version: "1.0"
paths:
...
  /v1/api/users/auth/google:
    post:
      consumes:
      - application/json
      description: Authenticate user with Google OAuth
      parameters:
      - description: Google auth data
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/user.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Authenticate with Google
      tags:
      - users/auth
  /v1/api/users/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/user.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Refresh access token
      tags:
      - users/auth
  /v1/api/users/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with Firebase authentication
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Create a new user
      tags:
      - users/auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"